chore: restructure repo for final release
- Archive AI scaffolding to _archive/ai-scaffolding/
- Archive legacy Kotlin project to _archive/ai-scaffolding/natasaku_kotlin/
- Move APK binaries to _releases/
- Clean root structure: source in natasaku_flutter/, docs/, LICENSE, standard files
- Fix 15 Dart analyzer issues and deprecation warnings
- Fix build configuration: compileSdk=36, targetSdk=34, minSdk=24
- Configure FileProvider and filepaths.xml resource for PDF exports
- Convert inline SharedPreferences keys to constants in AppConstants
- Build verified: flutter build apk --release exits 0
- APK: natasaku-v0.1.1-release.apk at _releases/

Version: 0.1.1+2
Min SDK: 24 (Android 7.0+)
Package: com.natasaku.app

+++ natasaku_flutter/lib/core/dev/mock_data_seeder.dart
@@ -59,7 +59,7 @@ class MockDataSeeder {
     txList.add({
       'id': 'tx_${idCounter++}',
       'date': DateTime(now.year, now.month, min(8, now.day)).toIso8601String(),
-      'amount': 300000.0,
+      'amount': 300000,
       'isExpense': false,
       'category': 'Pemasukan',
       'note': 'Freelance desain',
@@ -75,7 +75,7 @@ class MockDataSeeder {
         final noteList = notes[cat]!;
         
         // Nominal belanja bervariasi Rp10.000 - Rp80.000
-        final amt = (10000 + _rng.nextInt(70) * 1000).toDouble();
+        final amt = 10000 + _rng.nextInt(70) * 1000;
 
         txList.add({
           'id': 'tx_${idCounter++}',
